+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added support for `directed=False` (induced subgraph) neighbor sampling for homogeneous graphs on CPU. Currently does not support `disjoint=True` or `distributed=True` ([#690](https://github.com/pyg-team/pyg-lib/pull/690))
 - Added Linux `aarch64` manylinux wheel and CUDA Docker image builds ([#683](https://github.com/pyg-team/pyg-lib/pull/683))
 
 ### Changed